refactored AnimationControl to handle both layer and minified filter playerback
