+++ src/components/src/common/slider/slider.tsx
@@ -68,6 +68,8 @@ type SliderProps = {
   marks?: number[];
   classSet?: {[key: string]: boolean};
   disabled: boolean;
+  className?: string;
+  style?: object;
 };
 
 export default class Slider extends Component {
@@ -175,6 +177,7 @@ export default class Slider extends Component {
 
   render() {
     const {
+      className,
       classSet,
       disabled,
       isRanged,
@@ -183,7 +186,8 @@ export default class Slider extends Component {
       value1,
       vertical,
       sliderHandleWidth,
-      showTooltip
+      showTooltip,
+      style
     } = this.props;
     const value0 = !isRanged && minValue > 0 ? minValue : this.props.value0;
     const currValDelta = value1 - value0;
@@ -194,10 +198,11 @@ export default class Slider extends Component {
 
     return (
       <SliderWrapper
-        className={classnames('kg-slider', {...classSet, disabled})}
+        className={classnames('kg-slider', {...classSet, disabled}, className)}
         ref={this.ref}
         isRanged={isRanged}
         vertical={vertical}
+        style={style}
       >
         <StyledRangeSlider className="kg-range-slider" vertical={vertical} ref={this.track}>
           <SliderHandle
