+++ src/components/src/filters/index.ts
@@ -23,3 +23,5 @@ export {default as MultiSelectFilter} from './multi-select-filter';
 export {default as TimeRangeFilter} from './time-range-filter';
 export {default as RangeFilter} from './range-filter';
 export {default as PolygonFilter} from './polygon-filter';
+export {default as TimeWidget} from './time-widget';
+export {default as TimeWidgetTop} from './time-widget-top';
